Fix settings updating in server

+++ server/src/services/gitService.js
@@ -32,11 +32,24 @@ class GitService {
 
   async cloneRepository(repoName) {
     try {
+      const repoUrl = `https://github.com/${repoName}.git`;
+      if (!(await this.checkIfRepositoryExists(repoUrl))) {
+        throw new Error('Repository does not exist');
+      }
       await fs.remove('./repo');
-      await exec(`git clone https://github.com/${repoName}.git repo`);
+      await exec(`git clone ${repoUrl} repo`);
     } catch (error) {
       console.error('GitService.updateRepository error\n', error.stderr);
       throw new HttpError(`Cannot find ${repoName} repository`, 400, 'GIT_CANNOT_FIND_REPO');
+    }
+  }
+
+  async checkIfRepositoryExists(repoUrl) {
+    try {
+      await exec(`git ls-remote ${repoUrl}`);
+      return true;
+    } catch (err) {
+      return false;
     }
   }
